Refactor diagram menu
